services/analytics-service/requirements.txt
fastapi==0.115.0
psycopg[binary]==3.2.3
PyJWT==2.9.0
uvicorn[standard]==0.30.6
